+++ dear-imgui/src/render/draw_data.rs
@@ -288,7 +288,7 @@ impl<'a> Iterator for DrawCmdIterator<'a> {
                     cmd.ClipRect.z,
                     cmd.ClipRect.w,
                 ],
-                texture_id: TextureId::from(cmd.TexRef._TexID as usize),
+                texture_id: TextureId::from(cmd.TexRef._TexID),
                 vtx_offset: cmd.VtxOffset as usize,
                 idx_offset: cmd.IdxOffset as usize,
             };
